+++ CHANGES.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 - Breaking change: Change `db.cmp()` to return a Promise to deliver the result
+- Deprecated: on `schema.indexes`, in place of the index `key` property,
+    `keyPath` should be used.
 - Fix: Ensure there is a promise rejection for a bad schema callback or
     bad IDBKeyRange-related call
 - Fix: Actually implement documented chaining of (short) event handlers